requirements.txt
bcrypt==3.2.0
cffi==1.15.0
cryptography==35.0.0
paramiko==2.8.0
pycparser==2.21
PyNaCl==1.4.0
PyYAML==6.0
six==1.16.0
